+++ documentation/playground/src/molecules/DatepickerExamples.tsx
@@ -1,4 +1,4 @@
-import { Calendar } from '@a-ui/react';
+import { Calendar, Datepicker } from '@a-ui/react';
 
 import { useState } from 'react';
 
@@ -15,8 +15,23 @@ export function DatepickerExamples() {
         <Calendar
           value={date}
           onChange={setDate}
+          unavailableTo="2023-02-02"
+          unavailableFrom="2023-02-28"
           unavailable={[new Date('2023-02-24').toISOString(), new Date('2023-02-25').toISOString()]}
         />
+      </div>
+      <h2>Datepicker</h2>
+      <div className="u-margin">
+        <p className="u-margin-top">Without value</p>
+        <Datepicker
+          calendarProps={{
+            unavailable: [new Date(Date.now()).toISOString()],
+            unavailableTo: new Date(Date.now() - 5 * 24 * 60 * 60 * 1000).toISOString(),
+            unavailableFrom: new Date(Date.now() + 5 * 24 * 60 * 60 * 1000).toISOString()
+          }}
+        />
+        <p className="u-margin-top">With value</p>
+        <Datepicker format="dd-MM-yyyy" value={new Date(Date.now()).toISOString()} />
       </div>
     </div>
   );