+++ packages/antwerp-ui/react-components/src/lib/base/icon/Icon.types.ts
@@ -1,6 +1,13 @@
+import { KeyboardEventHandler, MouseEventHandler } from 'react';
+
 export interface IconProps {
   name: string;
   thin?: boolean;
+  role?: string;
+  className?: string;
   screenReaderText?: string;
   qa?: string;
+  tabIndex?: number;
+  onClick?: MouseEventHandler<HTMLSpanElement>;
+  onKeyDown?: KeyboardEventHandler<HTMLSpanElement>;
 }