+++ packages/antwerp-ui/react-components/src/lib/molecules/datepicker/index.ts
@@ -1,2 +1,3 @@
 export { Calendar } from './Calendar';
-export type { CalendarProps } from './datepicker.types';
+export { Datepicker } from './Datepicker';
+export type { CalendarProps, DatepickerProps } from './Datepicker.types';